pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "server-monitor-mcp"
version = "0.1.0"
description = "MCP server for remote server monitoring"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "大财主 & 发财一号", email = "example@domain.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=0.9.0",
    "paramiko>=3.4.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "colorlog>=6.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",
]

[project.scripts]
server-monitor-mcp = "server_monitor_mcp.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
